+++ app1/urls.py
@@ -4,6 +4,4 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('tables/', views.tables, name='tables'),
-    path('billing/', views.billing, name='billing'),
 ]