+++ app1/views.py
@@ -5,26 +5,5 @@ from django.http import HttpResponse
 
 def index(request):
 
-    # Just dummy text
-    # return HttpResponse("Hello, world. You're at the polls index.")
-    
     # Page from the theme 
-    # return render(request, 'pages/dashboard.html')
-
-    # Custom Page  
-    return render(request, 'custom.html')
-
-
-def tables(request):
-    # Page from the theme 
-    # return render(request, 'pages/tables.html')
-
-    # Custom Page  
-    return render(request, 'custom_table.html')
-
-def billing(request):
-    # Page from the theme 
-    # return render(request, 'pages/billing.html')
-
-    # Custom Page  
-    return render(request, 'custom_billing.html')
+    return render(request, 'pages/dashboard.html')