+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [1.0.1] 2022-11-20
+### Improvements
+
+- Update or [Django Admin Argon](https://github.com/app-generator/django-admin-argon-dashboard) `v1.0.7`
+- Integrate all pages for `Django.contrib.AUTH`
+- Integrate all theme links
+
 ## [1.0.0] 2022-11-10
 ### Initial Release